Add a switch to save assets using the referer url.
The `--trust-server-names` wget flag saves assets using the destination path of a redirect. This results in broken links when generating a static site.

The option has been added as a switch to maintain the default configuration and prevent breaking changes.

Note: from a file size perspective this is suboptimal as it results in duplicate images being generated.